+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Support tweaking theme style through `<style>` element or `style` global directive ([#25](https://github.com/marp-team/marpit/pull/25))
 * Add PostCSS import rollup plugin to work `@charset` and `@import` at-rules correctly ([#26](https://github.com/marp-team/marpit/pull/26))
 * Change role of pagination layer to pseudo layer on advanced background ([#27](https://github.com/marp-team/marpit/pull/27))
+* Fix over-stripped comments in the inline code ([#28](https://github.com/marp-team/marpit/pull/28))
 * Hide `section::after` pseudo-element without pagination ([#29](https://github.com/marp-team/marpit/pull/29))
 
 ## v0.0.5 - 2018-05-12